requirements.txt
fastapi
mysql-connector-python

uvicorn
sqlalchemy
passlib
bcrypt
PyJWT
python-jose
pymysql
python-multipart
pydantic
pandas
numpy
joblib
pyyaml
cryptography
aiofiles